→ Moves: tracking of pages card click to PagesCardViewModelSlice
