+++ WordPress/src/test/java/org/wordpress/android/ui/mysite/cards/dashboard/pages/PagesCardViewModelSliceTest.kt
@@ -16,6 +16,7 @@ import org.wordpress.android.ui.mysite.MySiteCardAndItemBuilderParams.PagesCardB
 import org.wordpress.android.ui.mysite.SelectedSiteRepository
 import org.wordpress.android.ui.mysite.SiteNavigationAction
 import org.wordpress.android.ui.mysite.cards.dashboard.CardsTracker
+import org.wordpress.android.ui.prefs.AppPrefsWrapper
 
 private const val MOCK_PAGE_ID = 1
 
@@ -28,6 +29,9 @@ class PagesCardViewModelSliceTest : BaseUnitTest() {
     @Mock
     lateinit var selectedSiteRepository: SelectedSiteRepository
 
+    @Mock
+    lateinit var appPrefsWrapper: AppPrefsWrapper
+
     private lateinit var pagesCardViewModelSlice: PagesCardViewModelSlice
 
     private lateinit var navigationActions: MutableList<SiteNavigationAction>
@@ -38,7 +42,8 @@ class PagesCardViewModelSliceTest : BaseUnitTest() {
     fun setUp() {
         pagesCardViewModelSlice = PagesCardViewModelSlice(
             cardsTracker,
-            selectedSiteRepository
+            selectedSiteRepository,
+            appPrefsWrapper
         )
         navigationActions = mutableListOf()
         pagesCardViewModelSlice.onNavigation.observeForever { event ->
@@ -68,7 +73,10 @@ class PagesCardViewModelSliceTest : BaseUnitTest() {
             pagesCardParams.onPagesItemClick.invoke(pagesParams)
 
             assertThat(navigationActions).containsOnly(SiteNavigationAction.OpenPagesDraftsTab(site, MOCK_PAGE_ID))
-            verify(cardsTracker).trackPagesItemClicked(PagesCardContentType.DRAFT)
+            verify(cardsTracker).trackCardItemClicked(
+                CardsTracker.Type.PAGES.label,
+                CardsTracker.PagesSubType.DRAFT.name
+            )
         }
 
     @Test
@@ -80,7 +88,10 @@ class PagesCardViewModelSliceTest : BaseUnitTest() {
             pagesCardParams.onPagesItemClick.invoke(pagesParams)
 
             assertThat(navigationActions).containsOnly(SiteNavigationAction.OpenPagesScheduledTab(site, MOCK_PAGE_ID))
-            verify(cardsTracker).trackPagesItemClicked(PagesCardContentType.SCHEDULED)
+            verify(cardsTracker).trackCardItemClicked(
+                CardsTracker.Type.PAGES.label,
+                CardsTracker.PagesSubType.SCHEDULED.name
+            )
         }
 
     @Test
@@ -92,7 +103,10 @@ class PagesCardViewModelSliceTest : BaseUnitTest() {
             pagesCardParams.onPagesItemClick.invoke(pagesParams)
 
             assertThat(navigationActions).containsOnly(SiteNavigationAction.OpenPages(site))
-            verify(cardsTracker).trackPagesItemClicked(PagesCardContentType.PUBLISH)
+            verify(cardsTracker).trackCardItemClicked(
+                CardsTracker.Type.PAGES.label,
+                CardsTracker.PagesSubType.PUBLISHED.name
+            )
         }
 
     @Test
